// src/pages/dashboard/components/type.ts
export interface PlayerData {
  name: string;
  age: string;
  gender: string;
  height: string;
  weight: string;
  club: string;
  position: string;
  nationality: string;
  dob: string;
  preferredFoot: string;
  playerNumber: string;
  email: string;
  phone: string;
  physicalEndurance: string;
  speed: string;
  strength: string;
  injuryHistory: string;
  matchesPlayed: string;
  minutesPlayed: string;
  goalsScored: string;
  assists: string;
  shotsOnTarget: string;
  passAccuracy: string;
  dribbles: string;
  tackles: string;
  interceptions: string;
  foulsCommitted: string;
  yellowCards: string;
  redCards: string;
  averageRating: string;
  trainingSessions: string;
  trainingHours: string;
  skillImprovement: string;
  coachFeedback: string;
  awards: string;
  trophies: string;
  records: string;
  education: string;
  languagesSpoken: string;
  hobbies: string;
  ballControl: string;
  passing: string;
  shooting: string;
  defending: string;
  tacticalAwareness: string;
  mentalToughness: string;
  leadership: string;
  teamwork: string;
  decisionMaking: string;
  socialMediaHandles: string;
  publicAppearances: string;
  fanEngagement: string;
  image: File | null;
  trainingVideos: File[];
  image_base64: string;
}